Fix the primary provider UI
In the 2fa setup screen, we show a radio button for the primary
provider, but we just use whatever is stored in the database. In other
places, we fallback to the first enabled provider if the one in the
database isn't active. This prevents inactive providers from being shown
as the primary provider.